# docs/openapi.yaml
openapi: 3.0.3
info:
  title: SchoolPilot API
  description: |
    Multi-tenant school management SaaS API.
    Authenticated routes require `Authorization: Bearer <accessToken>`.
    OpenAPI spec also served at `GET /api/docs/openapi.yaml`.
  version: 1.0.0
  contact:
    name: Great Goodluck Schools
servers:
  - url: http://localhost:4000
    description: Local development
  - url: https://api.yourschoolpilot.com
    description: Production
...
paths:
...
  /api/platform/plans:
    get:
      tags: [Platform]
      summary: List subscription plans (Super Admin)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Plan list with limits
    post:
      tags: [Platform]
      summary: Create subscription plan
      security: [{ bearerAuth: [] }]
      responses:
        '201':
          description: Created
  /api/platform/feature-flags:
    get:
      tags: [Platform]
      summary: List feature flags
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Feature flag list
  /api/platform/billing:
    get:
      tags: [Platform]
      summary: Billing overview
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Renewals, failed payments, coupons
  /api/platform/announcements:
    post:
      tags: [Platform]
      summary: Create platform announcement
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [title, body]
              properties:
                title: { type: string }
                body: { type: string }
                broadcast: { type: boolean }
                targetRoles:
                  type: array
                  items: { type: string }
      responses:
        '201':
          description: Created
...
  /api/support/tickets:
    get:
      tags: [Support]
      summary: List support tickets
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Ticket list
    post:
      tags: [Support]
      summary: Open support ticket
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [subject, body]
              properties:
                subject: { type: string }
                body: { type: string }
                category: { type: string }
      responses:
        '201':
          description: Created
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT